pyproject.toml
[project]
name = "aicoding"
version = "0.1.0"
description = "AI-Driven Data Science Problem Solver"
readme = "README.md"
authors = [
    { name = "Russell Poldrack", email = "poldrack@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=1.26,<3",
    "pandas>=2.1,<3",
    "scipy>=1.11,<2",
    "scikit-learn>=1.3,<2",
    "statsmodels>=0.14,<1",
    "matplotlib>=3.8,<4",
    "biopython>=1.83,<2",
    "nibabel>=5.2,<6",
    "radon>=6.0,<7",
    "networkx>=3.2,<4",
    "requests>=2.31,<3",
    "Pillow>=10.0,<12",
    "gensim>=4.3,<5",
    "nltk>=3.8,<4",
    "torch>=2.1",
    "markdown>=3.5,<4",
    "fpdf2>=2.7,<3",
    "textblob>=0.17,<1",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9",
    "pytest-cov>=4.1,<6",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
